More convenient API for inference of ModelHub models with ImageProxy input (#454)
* Extend the ModelHub's models with an API that accepts ImageProxy as an input.

* Now user not required to set targetRotation property by hand

The rotation operation is quite expensive, so it's beneficial to make a rotation after resizing, although it makes API less intuitive.

+++ onnx/src/androidMain/kotlin/org/jetbrains/kotlinx/dl/api/inference/onnx/CameraXCompatibleModel.kt
@@ -8,5 +8,5 @@ public interface CameraXCompatibleModel {
      * Target image rotation.
      * @see [ImageInfo](https://developer.android.com/reference/androidx/camera/core/ImageInfo)
      */
-    public var targetRotation: Float
+    public var targetRotation: Int
 }
